Correct user name bug and separation of TimeLag and ServerTime
