+++ Assets/_GameDevTwitchPlays12/Teams/Team Twitch/Module 1/Scripts/Mockup/Mockup_GameToServerTimeLag.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 83256afd13d28f44f96010f6424de146
+timeCreated: 1516352117
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
